doc_parser/requirements.txt
requests
PyYAML
huggingface-hub
PyPDF2
langchain
streamlit
text_generation
faiss-cpu 
tiktoken
docx2txt
networkx
openpyxl
python-pptx
odfpy